Shortcut factories instead of shortcuts registration.
The global (module-level) shortcut store was a bad idea and that
became apparent quickly.

Suppose a component registers a shortcut importing react-intl-ns
using the dist bundle and an application imports react-intl-ns/main.
You would then likely end up with with two versions of the module in
a bundle, with separate shortcut registries, the one available to the
application missing the component-registered shortcut.

Now integrating modules are responsible for exporting their shortcuts,
but have higher level helpers at their disposal.