Back to the @injectIntl pattern for string shortcuts.
The intlRef solution was a bit too fragile. It relied on the provided
function being called earlier than for normal refs (on namespace child
context determination). Moreover, the intl holder trick was not elegant
and needed too much explaining.

Decorating with @injectIntl or adding intl to contextTypes requires a
few more lines, but is clear and declarative.